pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pololoshka <sokolova.polya.98gmail.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
psycopg = {extras = ["binary"], version = "^3.1.17"}
channels = "^4.0.0"
channels-redis = "^4.2.0"
daphne = "^4.1.0"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"

[tool.poetry.group.dev.dependencies]
# Linters and Formatters
ruff = "*"
pre-commit = "*"
mypy = "*"
django-stubs = "*"
djangorestframework-stubs = "*"
pytest = "^8.0.0"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "src.settings"

[tool.ruff]
target-version = "py312"
extend-exclude = [".venv"]
line-length = 100

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/#error-e
select = [
  "E",   # Error
  "F",   # Pyflakes
  "I",   # Isort
  "N",   # pep8-naming
  "S",   # flake8-bandit
  "B",   # flake8-bugbear
  "C4",  # flake8-comprehensions
  "ICN", # flake8-import-conventions
  "PIE", # flake8-pie
  "PT",  # flake8-pytest-style
  "Q",   # flake8-quotes
  "SIM", # flake8-simplify
  "C90", # mccabe for complexicty
  "UP",  # pyupgrade
  "T20", # flake8-print
]
ignore = [
  "S311",  # Allow standard pseudo-random
]

[tool.ruff.lint.extend-per-file-ignores]
"test*" = [
  "S101", # Allow use of asserts
  "E501", # Unlimited line length
]
